% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShiny.R
\name{runShiny}
\alias{runShiny}
\title{A wrapper function to run Shiny Apps from \code{FDPDataAtlas}.}
\usage{
runShiny(app = "FDPDataAtlas")
}
\arguments{
\item{app}{FDPDataAtlas}
}
\value{
FDPDataAtlas shiny app
}
\description{
Running this function will launch the FDPDataAtlas shiny
}
